Refactoring: optimize and reorganize RenderSideParts
Use static lazy init by enum
