+++ src/main/scala/ru/mousecray/endmagic/blocks/decorative/polished/obsidian/IPolishedObsidian.java
@@ -1,11 +1,10 @@
 package ru.mousecray.endmagic.blocks.decorative.polished.obsidian;
 
 import net.minecraft.block.state.IBlockState;
-
-import java.util.function.Function;
+import ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSidePartsHolder.RenderSideParts;
 
 public interface IPolishedObsidian {
 
-    RenderSideParts2 getObsidianParts(IBlockState state);
+    RenderSideParts getObsidianParts(IBlockState state);
 
 }
