+++ src/main/scala/ru/mousecray/endmagic/blocks/decorative/polished/obsidian/Utils.java
@@ -4,7 +4,7 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
-import ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSideParts2.HorizontalFaceVisibility;
+import ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSidePartsHolder.RenderSideParts;
 
 import java.util.Arrays;
 import java.util.Map;
@@ -12,17 +12,16 @@ import java.util.function.Function;
 
 import static java.util.stream.Collectors.toMap;
 import static net.minecraft.block.BlockSlab.HALF;
-import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSideParts2.HorizontalFaceVisibility.*;
-import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSideParts2.PROPERTY;
+import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSidePartsHolder.RenderSideParts.FaceVisibility.*;
 
 public class Utils {
 
-    static RenderSideParts2 getObsidianSlabParts(IBlockState state) {
+    static RenderSideParts getObsidianSlabParts(IBlockState state) {
         switch (state.getValue(HALF)) {
             case TOP:
-                return RenderSideParts2.apply(visible_all, invisible_all, invisible_bottom, invisible_bottom, invisible_bottom, invisible_bottom);
+                return RenderSideParts.apply(visible_all, invisible_all, invisible_bottom, invisible_bottom, invisible_bottom, invisible_bottom);
             case BOTTOM:
-                return RenderSideParts2.apply(invisible_all, visible_all, invisible_top, invisible_top, invisible_top, invisible_top);
+                return RenderSideParts.apply(invisible_all, visible_all, invisible_top, invisible_top, invisible_top, invisible_top);
             default:
                 throw new IllegalArgumentException("Unsupported value of HALF property");
         }
@@ -30,7 +29,7 @@ public class Utils {
 
     static IBlockState getActualObsidianSlabState(IBlockState state, IBlockAccess worldIn, BlockPos pos) {
 
-        Map<EnumFacing, HorizontalFaceVisibility> actualVisibility = getActualVisibilityMap(worldIn, pos);
+        Map<EnumFacing, RenderSideParts.FaceVisibility> actualVisibility = getActualVisibilityMap(worldIn, pos);
 
         switch (state.getValue(HALF)) {
             case TOP:
@@ -45,14 +44,14 @@ public class Utils {
                 throw new IllegalArgumentException("Unsupported value of HALF property");
         }
 
-        return state.withProperty(PROPERTY, RenderSideParts2.apply(actualVisibility));
+        return state.withProperty(RenderSideParts.PROPERTY, RenderSideParts.apply(actualVisibility));
     }
 
     static IBlockState getActualObsidianFullState(IBlockState state, IBlockAccess worldIn, BlockPos pos) {
-        return state.withProperty(PROPERTY, RenderSideParts2.apply(getActualVisibilityMap(worldIn, pos)));
+        return state.withProperty(RenderSideParts.PROPERTY, RenderSideParts.apply(getActualVisibilityMap(worldIn, pos)));
     }
 
-    private static Map<EnumFacing, HorizontalFaceVisibility> getActualVisibilityMap(IBlockAccess worldIn, BlockPos pos) {
+    private static Map<EnumFacing, RenderSideParts.FaceVisibility> getActualVisibilityMap(IBlockAccess worldIn, BlockPos pos) {
         return Arrays.stream(EnumFacing.values())
                 .collect(toMap(Function.identity(),
                         facing -> getActualVisibility(
@@ -61,10 +60,10 @@ public class Utils {
                         )));
     }
 
-    private static HorizontalFaceVisibility getActualVisibility(IBlockState state, EnumFacing side) {
+    private static RenderSideParts.FaceVisibility getActualVisibility(IBlockState state, EnumFacing side) {
         if (state.getBlock() instanceof IPolishedObsidian) {
-            RenderSideParts2 obsidianParts = ((IPolishedObsidian) state.getBlock()).getObsidianParts(state);
-            HorizontalFaceVisibility existedObsidianPart = obsidianParts.get(side);
+            RenderSideParts obsidianParts = ((IPolishedObsidian) state.getBlock()).getObsidianParts(state);
+            RenderSideParts.FaceVisibility existedObsidianPart = obsidianParts.get(side);
             return existedObsidianPart.invert();
         } else
             return visible_all;
