+++ src/main/scala/ru/mousecray/endmagic/client/render/model/baked/TranslucentPartsModel.java
@@ -4,15 +4,15 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.block.model.BakedQuad;
 import net.minecraft.client.renderer.block.model.IBakedModel;
 import net.minecraft.util.EnumFacing;
-import ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSideParts2;
+import ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSidePartsHolder.RenderSideParts;
 
 import javax.annotation.Nullable;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSideParts2.HorizontalFaceVisibility.invisible_bottom;
-import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSideParts2.HorizontalFaceVisibility.invisible_top;
+import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSidePartsHolder.RenderSideParts.FaceVisibility.invisible_bottom;
+import static ru.mousecray.endmagic.blocks.decorative.polished.obsidian.RenderSidePartsHolder.RenderSideParts.FaceVisibility.invisible_top;
 
 public class TranslucentPartsModel extends BakedModelDelegate {
     public TranslucentPartsModel(IBakedModel base) {
@@ -23,7 +23,7 @@ public class TranslucentPartsModel extends BakedModelDelegate {
     public List<BakedQuad> getQuads(@Nullable IBlockState state, @Nullable EnumFacing side, long rand) {
         List<BakedQuad> quads = super.getQuads(state, side, rand);
         if (state != null) {
-            RenderSideParts2 renderSideParts = state.getValue(RenderSideParts2.PROPERTY);
+            RenderSideParts renderSideParts = state.getValue(RenderSideParts.PROPERTY);
             return quads.stream().flatMap(q -> {
                 if (q instanceof MarkedBakedQuad) {
                     String side_part = ((MarkedBakedQuad) q).customValues.getOrDefault("side_part", "").toString();
@@ -39,7 +39,7 @@ public class TranslucentPartsModel extends BakedModelDelegate {
             return quads;
     }
 
-    private boolean canRenderQuad(BakedQuad q, String side_part, RenderSideParts2 renderSideParts) {
+    private boolean canRenderQuad(BakedQuad q, String side_part, RenderSideParts renderSideParts) {
         switch (q.getFace()) {
             case DOWN:
                 return renderSideParts.down.toString().equals(side_part);
@@ -58,7 +58,7 @@ public class TranslucentPartsModel extends BakedModelDelegate {
         }
     }
 
-    private boolean isSidePartRelevant(String side_part, RenderSideParts2.HorizontalFaceVisibility actualVisibility) {
+    private boolean isSidePartRelevant(String side_part, RenderSideParts.FaceVisibility actualVisibility) {
         switch (actualVisibility) {
             case visible_all:
                 return true;
